requirements.txt
# Python dependencies for RRC Refinery Pipeline

# Core dependencies
databricks-sdk>=0.18.0
anthropic>=0.18.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data processing
pandas>=2.0.0

# SQL parsing (for validation)
sqlparse>=0.4.4

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
